+++ src/data/action_creators.js
@@ -4,5 +4,12 @@ const ActionCreators = {
       type: "TABGROUPS_RECEIVE",
       tabgroups: tabgroups
     };
+  },
+
+  setGroupCloseTimeout: function(timeout) {
+    return {
+      type: "GROUP_CLOSE_TIMEOUT_RECIEVE",
+      closeTimeout: timeout
+    };
   }
 };