+++ src/data/components/groupcontrols.js
@@ -5,13 +5,14 @@ const GroupControls = React.createClass({
     onEdit: React.PropTypes.func,
     onEditAbort: React.PropTypes.func,
     onEditSave: React.PropTypes.func,
-    onExpand: React.PropTypes.func
+    onExpand: React.PropTypes.func,
+    onUndoCloseClick: React.PropTypes.func
   },
 
-  render: function() {
-    let editControls;
+  getEditControls: function() {
+    let controls;
     if (this.props.editing) {
-      editControls = [
+      controls = [
         React.DOM.i({
           className: "group-edit fa fa-fw fa-check",
           onClick: this.props.onEditSave
@@ -22,10 +23,30 @@ const GroupControls = React.createClass({
         })
       ];
     } else {
-      editControls = React.DOM.i({
+      controls = React.DOM.i({
         className: "group-edit fa fa-fw fa-pencil",
         onClick: this.props.onEdit
       });
+    }
+
+    return controls;
+  },
+
+  getClosingControls: function() {
+    return [
+      React.DOM.i({
+        className: "group-close-undo fa fa-fw fa-undo",
+        onClick: this.props.onUndoCloseClick
+      })
+    ];
+  },
+
+  render: function() {
+    let groupControls;
+    if (this.props.closing) {
+      groupControls = this.getClosingControls();
+    } else {
+      groupControls = this.getEditControls();
     }
 
     let expanderClasses = classNames({
@@ -40,7 +61,7 @@ const GroupControls = React.createClass({
       {
         className: "group-controls"
       },
-      editControls,
+      groupControls,
       React.DOM.i({
         className: "group-close fa fa-fw fa-times",
         onClick: this.props.onClose